+++ packages/react-server/examples/next/app/test/jsx-in-js/_jsx-in-client.js
